% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_spatial_utils.R
\name{compute_distance}
\alias{compute_distance}
\title{Compute Euclidean distance matrix}
\usage{
compute_distance(coords1, coords2)
}
\arguments{
\item{coords1}{Matrix of first set of coordinates (n x 2)}

\item{coords2}{Matrix of second set of coordinates (m x 2)}
}
\value{
Matrix of distances (n x m)
}
\description{
Compute Euclidean distance matrix
}
