% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_spatial_utils.R
\name{compute_ess}
\alias{compute_ess}
\title{Calculate Effective Sample Size}
\usage{
compute_ess(weights)
}
\arguments{
\item{weights}{Numeric vector of weights}
}
\value{
Numeric ESS value
}
\description{
Calculate Effective Sample Size
}
