% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_spatial_utils.R
\name{get_weights}
\alias{get_weights}
\title{Get Kernel Weights}
\usage{
get_weights(dists, bandwidth, kernel_type = "bisquare")
}
\arguments{
\item{dists}{Distance matrix}

\item{bandwidth}{Bandwidth value}

\item{kernel_type}{Kernel name}
}
\value{
Weight matrix
}
\description{
Get Kernel Weights
}
