% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_spatial_utils.R
\name{kernel_bisquare}
\alias{kernel_bisquare}
\title{Kernel functions}
\usage{
kernel_bisquare(d, h)
}
\arguments{
\item{d}{Vector or matrix of distances}

\item{h}{Bandwidth}
}
\value{
Vector or matrix of weights
}
\description{
Kernel functions
}
