% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_diagnostics.R
\name{plot.gw_causal_forest}
\alias{plot.gw_causal_forest}
\title{Plot method for gw_causal_forest}
\usage{
\method{plot}{gw_causal_forest}(
  x,
  type = c("effect_map", "ess_map", "cv_curve", "overlap"),
  newX = NULL,
  newcoords = NULL,
  ...
)
}
\arguments{
\item{x}{A gw_causal_forest object.}

\item{type}{Character, plot type: "effect_map", "ess_map", "cv_curve", or "overlap".}

\item{newX}{Matrix of covariates (required for effect_map).}

\item{newcoords}{Matrix of coordinates (required for effect_map).}

\item{...}{Additional arguments (unused).}
}
\value{
A ggplot2 object.
}
\description{
Plot method for gw_causal_forest
}
