% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_data_prep.R
\name{print.gwcf_data}
\alias{print.gwcf_data}
\title{Print method for gwcf_data}
\usage{
\method{print}{gwcf_data}(x, ...)
}
\arguments{
\item{x}{A gwcf_data object.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns x.
}
\description{
Print method for gwcf_data
}
