% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_spatial_utils.R
\name{select_anchors}
\alias{select_anchors}
\title{Select Anchors using K-means}
\usage{
select_anchors(coords, n_anchors, seed = NULL)
}
\arguments{
\item{coords}{Coordinate matrix}

\item{n_anchors}{Number of anchors}

\item{seed}{Random seed}
}
\value{
Matrix of anchor coordinates
}
\description{
Select Anchors using K-means
}
