% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_inference.R
\name{spatial_bootstrap_with_data}
\alias{spatial_bootstrap_with_data}
\title{Spatial Bootstrap with Data provided}
\usage{
spatial_bootstrap_with_data(
  object,
  Y_train,
  W_train,
  X_train,
  coords_train,
  newX,
  newcoords,
  n_bootstrap = 200,
  n_blocks = 10,
  conf_level = 0.95,
  seed = NULL,
  verbose = TRUE
)
}
\arguments{
\item{object}{Fitted model.}

\item{Y_train}{Training outcome.}

\item{W_train}{Training treatment.}

\item{X_train}{Training covariates.}

\item{coords_train}{Training coordinates.}

\item{newX}{Prediction covariates.}

\item{newcoords}{Prediction coordinates.}

\item{n_bootstrap}{Number of bootstrap iterations.}

\item{n_blocks}{Number of spatial blocks.}

\item{conf_level}{Confidence level (default 0.95).}

\item{seed}{Random seed.}

\item{verbose}{Logical, whether to print progress.}
}
\description{
Spatial Bootstrap with Data provided
}
